feat: 공통 모달 타이틀·서브타이틀 변경 기능 구현 수정 1) bindModalEvents 내부에 updateModalText 함수 코드 이동
